example_project/requirements.txt
pytest==5.4.1
pytest-cov==2.8.1
pylint==2.4.4
djangorestframework==3.11.0
drf-writable-nested==0.5.4
../dist/djangorestframework-jsonschema-0.1.0.tar.gz
